+++ devenv.nix
@@ -1,0 +1,9 @@
+{ pkgs, ... }:
+
+{
+  packages = [
+    pkgs.nodejs_20
+  ];
+
+  languages.typescript.enable = true;
+}