+++ src/test/suite/commands/commands.test.ts
@@ -6,19 +6,13 @@ import vscode from 'vscode'
 import { workspaceRoot } from '.'
 
 context('commands in the test workspace', function () {
-	const file = path.join(workspaceRoot, '.envrc')
+	const file = path.join(workspaceRoot, 'devenv.nix')
 
 	context('without open editors', function () {
-		specify('direnv.open opens the .envrc file', async function () {
-			await vscode.commands.executeCommand('direnv.open')
-			const path = vscode.window.activeTextEditor?.document.fileName
-			assert.equal(path, file)
-		})
-
-		specify('direnv.create opens the existing .envrc file', async function () {
-			await vscode.commands.executeCommand('direnv.create')
-			const path = vscode.window.activeTextEditor?.document.fileName
-			assert.equal(path, file)
+		specify('devenv.open opens the devenv.nix file', async function () {
+			await vscode.commands.executeCommand('devenv.open')
+			const filePath = vscode.window.activeTextEditor?.document.fileName
+			assert.equal(filePath, file)
 		})
 	})
 
@@ -28,16 +22,10 @@ context('commands in the test workspace', function () {
 			await vscode.commands.executeCommand('vscode.open', vscode.Uri.file(text))
 		})
 
-		specify('direnv.open switches to the .envrc file', async function () {
-			await vscode.commands.executeCommand('direnv.open')
-			const path = vscode.window.activeTextEditor?.document.fileName
-			assert.equal(path, file)
-		})
-
-		specify('direnv.create switches to the existing .envrc file', async function () {
-			await vscode.commands.executeCommand('direnv.create')
-			const path = vscode.window.activeTextEditor?.document.fileName
-			assert.equal(path, file)
+		specify('devenv.open switches to the devenv.nix file', async function () {
+			await vscode.commands.executeCommand('devenv.open')
+			const filePath = vscode.window.activeTextEditor?.document.fileName
+			assert.equal(filePath, file)
 		})
 	})
 })