+++ src/test/suite/commands/subdir/commands.test.ts
@@ -6,20 +6,13 @@ import vscode from 'vscode'
 import { workspaceRoot } from '.'
 
 context('commands in the subdirectory workspace', function () {
-	const file = path.join(workspaceRoot, '../.envrc')
-	const subfile = path.join(workspaceRoot, '.envrc')
+	const file = path.join(workspaceRoot, 'devenv.nix')
 
 	context('without open editors', function () {
-		specify('direnv.open opens the parent .envrc file', async function () {
-			await vscode.commands.executeCommand('direnv.open')
-			const path = vscode.window.activeTextEditor?.document.fileName
-			assert.equal(path, file)
-		})
-
-		specify('direnv.create opens a new .envrc file in the subdirectory', async function () {
-			await vscode.commands.executeCommand('direnv.create')
-			const path = vscode.window.activeTextEditor?.document.fileName
-			assert.equal(path, subfile)
+		specify('devenv.open opens the devenv.nix file', async function () {
+			await vscode.commands.executeCommand('devenv.open')
+			const filePath = vscode.window.activeTextEditor?.document.fileName
+			assert.equal(filePath, file)
 		})
 	})
 })