+++ src/test/suite/interface/index.ts
@@ -1,9 +1,0 @@
-import path from 'path'
-
-import { runSuite } from '..'
-
-export const workspaceRoot = path.resolve(__dirname, '../../../../test/workspace')
-
-export async function run() {
-	await runSuite(workspaceRoot, __dirname)
-}